--- a/src/test/compatibility.ts
+++ b/src/test/compatibility.ts
@@ -14,7 +14,7 @@ import {
   plugin as tseslintBottomPlugin,
 } from 'typescript-eslint_bottom'
 import eslintCommentsBottomPlugin from 'eslint-plugin-eslint-comments_bottom'
-import importBottomPlugin from 'eslint-plugin-import_bottom'
+import importXBottomPlugin from 'eslint-plugin-import-x_bottom'
 import nBottomPlugin from 'eslint-plugin-n_bottom'
 import promiseBottomPlugin from 'eslint-plugin-promise_bottom'
 import _ from 'lodash'
@@ -23,7 +23,7 @@ test('bottom dep version is minimum of dep range', (t) => {
   const bottomDepsThatAreNotMinOfDepRange = [
     ['typescript-eslint', ourDeps] as const,
     ['eslint', ourPeerDeps] as const,
-    ['eslint-plugin-import', ourDeps] as const,
+    ['eslint-plugin-import-x', ourDeps] as const,
     ['eslint-plugin-n', ourDeps] as const,
     ['eslint-plugin-promise', ourDeps] as const,
   ]
@@ -60,7 +60,7 @@ test('our configuration is compatible with the plugins and parser at bottom of d
     plugins: {
       '@typescript-eslint': tseslintBottomPlugin,
       'eslint-comments': eslintCommentsBottomPlugin,
-      import: importBottomPlugin,
+      'import-x': importXBottomPlugin,
       n: nBottomPlugin,
       promise: promiseBottomPlugin,
     },