--- a/src/test/expected-exported-value/_index.ts
+++ b/src/test/expected-exported-value/_index.ts
@@ -1,12 +1,12 @@
 import type { TSESLint } from '@typescript-eslint/utils'
-import { parser, plugin as tseslintPlugin } from 'typescript-eslint'
 import eslintCommentsPlugin from 'eslint-plugin-eslint-comments'
-import importPlugin from 'eslint-plugin-import'
+import importXPlugin from 'eslint-plugin-import-x'
 import nPlugin from 'eslint-plugin-n'
 import promisePlugin from 'eslint-plugin-promise'
+import { parser, plugin as tseslintPlugin } from 'typescript-eslint'
 import { expectedEslintCommentsRules } from './_eslint-comments.js'
 import { expectedEslintRules } from './_eslint.js'
-import { expectedImportRules } from './_import.js'
+import { expectedImportXRules } from './_import-x.js'
 import { expectedNRules } from './_n.js'
 import { expectedPromiseRules } from './_promise.js'
 import { expectedTseslintRules } from './_typescript-eslint.js'
@@ -14,7 +14,7 @@ import { expectedTseslintRules } from './_typescript-eslint.js'
 export const expectedExportedRules = {
   ...expectedEslintCommentsRules,
   ...expectedEslintRules,
-  ...expectedImportRules,
+  ...expectedImportXRules,
   ...expectedNRules,
   ...expectedPromiseRules,
   ...expectedTseslintRules,
@@ -35,7 +35,7 @@ export const expectedExportedValue: TSESLint.FlatConfig.Config = {
   plugins: {
     '@typescript-eslint': tseslintPlugin,
     'eslint-comments': eslintCommentsPlugin,
-    import: importPlugin,
+    'import-x': importXPlugin,
     n: nPlugin,
     promise: promisePlugin,
   },