--- a/src/test/rules.ts
+++ b/src/test/rules.ts
@@ -2,7 +2,7 @@ import test from 'ava'
 import { plugin as pluginTseslint } from 'typescript-eslint'
 import pluginEslintComments from 'eslint-plugin-eslint-comments'
 import pluginN from 'eslint-plugin-n'
-import pluginImport from 'eslint-plugin-import'
+import pluginImportX from 'eslint-plugin-import-x'
 import pluginPromise from 'eslint-plugin-promise'
 import { equivalents, ourRules } from './_util.js'
 import _ from 'lodash'
@@ -12,7 +12,7 @@ import { rulesToConsider } from './_rules_to_consider.js'
 import { rulesPerPlugin } from '../plugin-usage.js'
 import { expectedEslintCommentsRules } from './expected-exported-value/_eslint-comments.js'
 import { expectedEslintRules } from './expected-exported-value/_eslint.js'
-import { expectedImportRules } from './expected-exported-value/_import.js'
+import { expectedImportXRules } from './expected-exported-value/_import-x.js'
 import { expectedNRules } from './expected-exported-value/_n.js'
 import { expectedPromiseRules } from './expected-exported-value/_promise.js'
 import { expectedTseslintRules } from './expected-exported-value/_typescript-eslint.js'
@@ -23,7 +23,8 @@ const knownEslintRules = new TSESLint.Linter({
 
 if (pluginEslintComments.rules === undefined) throw new Error()
 if (pluginN.rules === undefined) throw new Error()
-if (pluginImport.rules === undefined) throw new Error()
+// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- intended
+if (pluginImportX.rules === undefined) throw new Error()
 if (pluginPromise.rules === undefined) throw new Error()
 if (pluginTseslint.rules === undefined) throw new Error()
 
@@ -31,7 +32,11 @@ const rulesets: Array<[TSESLint.Linter.Plugin, string]> = [
   [pluginEslintComments.rules, 'eslint-comments'],
   [pluginTseslint.rules, '@typescript-eslint'],
   [pluginN.rules, 'n'],
-  [pluginImport.rules, 'import'],
+  [
+    // @ts-expect-error -- too strict...
+    pluginImportX.rules,
+    'import-x',
+  ],
   [pluginPromise.rules, 'promise'],
 ]
 
@@ -140,7 +145,7 @@ test('rule lists and objects are sorted', (t) => {
     ),
     'expected rules/eslint-comments': Object.keys(expectedEslintCommentsRules),
     'expected rules/eslint': Object.keys(expectedEslintRules),
-    'expected rules/import': Object.keys(expectedImportRules),
+    'expected rules/import-x': Object.keys(expectedImportXRules),
     'expected rules/n': Object.keys(expectedNRules),
     'expected rules/promise': Object.keys(expectedPromiseRules),
     'expected rules/@typescript-eslint': Object.keys(expectedTseslintRules),