add `eslint-import-resolver-typescript`

---

Just a proposal, not sure how to test `t.deepEqual(exported, expectedExportedValue)` here because a function is added for https://github.com/import-js/eslint-import-resolver-typescript#eslintconfigjs

Wait for feedback first.

cc @mightyiam