@JounQin 2025-04-12
> Hey @JounQin. Thank you for your effort. Please see [#1509 (comment)](https://github.com/mightyiam/eslint-config-love/pull/1509#issuecomment-2183846005).

OK, I'll remove `eslint-import-resolver-typescript` then.