@JounQin 2025-04-12
> What's the motivation for replacing `import` with `import-x`, please?

Just quote from another PR: https://github.com/shlinkio/js-coding-standard/pull/129#issuecomment-2764434829