@mightyiam 2025-04-15
So this is about performance? Do you know how much performance?