@JounQin 2025-04-15
> So this is about performance? Do you know how much performance?

Not only about performance but also dependencies chain, see also https://e18e.dev/guide/cleanup.html, but I ran for you on your this simple repo, it should be more efficient on real larger projects:

Before:

```console
 eslint-config-love on  main ❯ TIMING=1 npmR lint

> eslint-config-love@119.0.0 lint
> eslint .

Rule                                        | Time (ms) | Relative
:-------------------------------------------|----------:|--------:
@typescript-eslint/no-unsafe-assignment     |   183.333 |    42.5%
@typescript-eslint/no-misused-promises      |    39.310 |     9.1%
import/no-duplicates                        |    35.481 |     8.2%
@typescript-eslint/no-floating-promises     |    22.237 |     5.2%
@typescript-eslint/naming-convention        |    15.134 |     3.5%
@typescript-eslint/no-deprecated            |    13.462 |     3.1%
@typescript-eslint/no-unsafe-return         |     7.101 |     1.6%
no-this-before-super                        |     6.263 |     1.5%
no-dupe-keys                                |     4.212 |     1.0%
@typescript-eslint/no-unnecessary-condition |     4.180 |     1.0%
```

After:

```console
 eslint-config-love on  refactor/import-x ❯ TIMING=1 npmR lint

> eslint-config-love@119.0.0 lint
> eslint .

Rule                                        | Time (ms) | Relative
:-------------------------------------------|----------:|--------:
@typescript-eslint/no-unsafe-assignment     |   193.841 |    45.0%
@typescript-eslint/no-misused-promises      |    44.112 |    10.2%
import-x/no-duplicates                      |    25.126 |     5.8%
@typescript-eslint/no-floating-promises     |    23.344 |     5.4%
@typescript-eslint/no-deprecated            |    14.252 |     3.3%
@typescript-eslint/naming-convention        |    13.621 |     3.2%
@typescript-eslint/no-unsafe-return         |     6.826 |     1.6%
@typescript-eslint/no-unnecessary-condition |     4.319 |     1.0%
@typescript-eslint/no-unsafe-argument       |     4.084 |     0.9%
@typescript-eslint/unbound-method           |     3.008 |     0.7%
```