@JounQin 2025-04-15
Sure, that depends on you, feel free to close if you don't like it.

`eslint-plugin-import` is not abandoned, just has much long dependencies chain, see also https://github.com/SukkaW/nolyfill